+++ PythonAPI/source/libcarla/Map.cpp
@@ -47,6 +47,12 @@ static auto GetTopology(const carla::client::Map &self) {
   return result;
 }
 
+static carla::geom::GeoLocation ToGeolocation(
+    const carla::client::Map &self,
+    const carla::geom::Location &location) {
+  return self.GetGeoReference().Transform(location);
+}
+
 void export_map() {
   using namespace boost::python;
   namespace cc = carla::client;
@@ -58,6 +64,7 @@ void export_map() {
     .def("get_waypoint", &cc::Map::GetWaypoint, (arg("location"), arg("project_to_road")=true))
     .def("get_topology", &GetTopology)
     .def("generate_waypoints", CALL_RETURNING_LIST_1(cc::Map, GenerateWaypoints, double), (args("distance")))
+    .def("transform_to_geolocation", &ToGeolocation, (arg("location")))
     .def("to_opendrive", CALL_RETURNING_COPY(cc::Map, GetOpenDrive))
     .def("save_to_disk", &SaveOpenDriveToDisk, (arg("path")=""))
     .def(self_ns::str(self_ns::self))